PickUp locations working as intended :)
